+++ src/agoradatatools/etl/transform/model_details.py
@@ -13,7 +13,7 @@ from agoradatatools.etl.utils import (
     normalize_null_values,
     delim_string_to_list,
 )
-from agoradatatools.etl.transform.model_ad_transform_utils import (
+from agoradatatools.etl.transform.transform_utils.model_ad_transform_utils import (
     build_transcriptomics_url,
     process_genetic_info,
     zero_pad_jax_ids,
