+++ tests/transform/test_model_ad_transform_utils.py
@@ -6,7 +6,7 @@ import pandas as pd
 import numpy as np
 import pytest
 
-from agoradatatools.etl.transform.model_ad_transform_utils import (
+from agoradatatools.etl.transform.transform_utils.model_ad_transform_utils import (
     build_transcriptomics_url,
     process_genetic_info,
     zero_pad_jax_ids,
